healthpilot-backend/requirements/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
cachetools==6.1.0
certifi==2025.7.14
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorlog==6.9.0
cssselect==1.3.0
cssutils==2.11.1
dataclasses-json==0.6.7
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
einops==0.8.1
et_xmlfile==2.0.0
fastapi==0.116.1
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
ftfy==6.3.1
gotrue==2.12.3
GPUtil==1.4.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
hyperframe==6.1.0
idna==3.10
imagesize==1.4.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langsmith==0.4.9
lxml==6.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
more-itertools==10.7.0
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.2
openai==1.98.0
opencv-contrib-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.11.1
packaging==25.0
paddleocr==3.1.0
paddlepaddle==3.1.0
paddlex==3.1.3
pandas==2.3.1
pdf2image==1.17.0
pillow==11.3.0
postgrest==1.1.1
premailer==3.10.0
prettytable==3.16.0
propcache==0.3.2
protobuf==6.31.1
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
realtime==2.6.0
redis==6.2.0
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rq==2.4.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
starlette==0.47.2
storage3==0.12.0
StrEnum==0.4.15
supabase==2.17.0
supafunc==0.10.1
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.4
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
websockets==15.0.1
wheel==0.45.1
yarl==1.20.1
zstandard==0.23.0
